+++ moli-release/releases/handlebars-helpers/var.js
@@ -1,4 +1,3 @@
 module.exports = function (name, value) {
-  console.log(name, value);
   this[name] = value;
 };